LayoutTests/http/tests/security/cross-frame-access-selection-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/foreign-domain-iframe-for-selection-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-selection.html. Domains, protocols and ports must match.

PASS: Cross-site access denied.
